You can now do 64bit ops

+++ src/test/kotlin/venusbackend/riscv/InstructionTest.kt
@@ -9,12 +9,12 @@ class InstructionTest {
     @Test
     fun correctOpcode() {
         val mcode: MachineCode = MachineCode(0x1ead12aa)
-        assertEquals(0x2a, mcode[InstructionField.OPCODE])
+        assertEquals(0x2a, mcode[InstructionField.OPCODE].toInt())
     }
 
     @Test fun setGetFields() {
         val mcode: MachineCode = MachineCode(0x1ead12aa)
         mcode[InstructionField.OPCODE] = 0x1b
-        assertEquals(0x1b, mcode[InstructionField.OPCODE])
+        assertEquals(0x1b, mcode[InstructionField.OPCODE].toInt())
     }
 }
