+++ src/test/kotlin/venusbackend/simulator/StaticDataTest.kt
@@ -38,10 +38,10 @@ class StaticDataTest {
         val linked = Linker.link(PandL)
         val sim = Simulator(linked)
         val offset = MemorySegments.STATIC_BEGIN
-        assertEquals('a'.toInt(), sim.loadByte(offset))
-        assertEquals(0, sim.loadByte(offset + 1))
-        assertEquals('b'.toInt(), sim.loadByte(offset + 2))
-        assertEquals(0, sim.loadByte(offset + 3))
+        assertEquals('a'.toInt(), sim.loadByte(offset).toInt())
+        assertEquals(0, sim.loadByte(offset + 1).toInt())
+        assertEquals('b'.toInt(), sim.loadByte(offset + 2).toInt())
+        assertEquals(0, sim.loadByte(offset + 3).toInt())
     }
 
     @Test fun linkedStaticBytes() {
